Extend Fieldset instead of creating Form
