+++ Showroom/examples/components.js
@@ -6,7 +6,7 @@ export ContextMenu from '../../ContextMenu/example'
 export Dialog from '../../Dialog/example'
 export Dropdown from '../../Dropdown/example'
 export Field from '../../Field/example'
-export Form from '../../Form/example'
+export Fieldset from '../../Fieldset/example'
 export Icons from '../../icons/example'
 export Input from '../../Input/example'
 export Installments from '../../Installments/example'
